move search in a separate e2e folder
