+++ e2e/search/components/search-number-range.e2e.ts
@@ -15,22 +15,22 @@
  * limitations under the License.
  */
 
-import { LoginPage } from '../../../pages/adf/loginPage';
-import { SearchDialog } from '../../../pages/adf/dialog/searchDialog';
-import DataTablePage = require('../../../pages/adf/dataTablePage');
-import { SearchResultsPage } from '../../../pages/adf/searchResultsPage';
-import { NavigationBarPage } from '../../../pages/adf/navigationBarPage';
-import { ConfigEditorPage } from '../../../pages/adf/configEditorPage';
-import { SearchFiltersPage } from '../../../pages/adf/searchFiltersPage';
+import { LoginPage } from '../../pages/adf/loginPage';
+import { SearchDialog } from '../../pages/adf/dialog/searchDialog';
+import DataTablePage = require('../../pages/adf/dataTablePage');
+import { SearchResultsPage } from '../../pages/adf/searchResultsPage';
+import { NavigationBarPage } from '../../pages/adf/navigationBarPage';
+import { ConfigEditorPage } from '../../pages/adf/configEditorPage';
+import { SearchFiltersPage } from '../../pages/adf/searchFiltersPage';
 
-import TestConfig = require('../../../test.config');
+import TestConfig = require('../../test.config');
 
 import AlfrescoApi = require('alfresco-js-api-node');
-import { UploadActions } from '../../../actions/ACS/upload.actions';
-import AcsUserModel = require('../../../models/ACS/acsUserModel');
-import FileModel = require('../../../models/ACS/fileModel');
+import { UploadActions } from '../../actions/ACS/upload.actions';
+import AcsUserModel = require('../../models/ACS/acsUserModel');
+import FileModel = require('../../models/ACS/fileModel');
 import { browser } from 'protractor';
-import resources = require('../../../util/resources');
+import resources = require('../../util/resources');
 import { SearchConfiguration } from '../search.config';
 
 describe('Search Number Range Filter', () => {
