+++ e2e/search/search_component.e2e.ts
@@ -17,21 +17,21 @@
 
 import { browser } from 'protractor';
 
-import { LoginPage } from '../../pages/adf/loginPage';
-import { SearchDialog } from '../../pages/adf/dialog/searchDialog';
-import { ContentServicesPage } from '../../pages/adf/contentServicesPage';
-import filePreviewPage = require('../../pages/adf/filePreviewPage');
-import { SearchResultsPage } from '../../pages/adf/searchResultsPage';
+import { LoginPage } from '../pages/adf/loginPage';
+import { SearchDialog } from '../pages/adf/dialog/searchDialog';
+import { ContentServicesPage } from '../pages/adf/contentServicesPage';
+import filePreviewPage = require('../pages/adf/filePreviewPage');
+import { SearchResultsPage } from '../pages/adf/searchResultsPage';
 
-import AcsUserModel = require('../../models/ACS/acsUserModel');
-import FileModel = require('../../models/ACS/fileModel');
-import FolderModel = require('../../models/ACS/folderModel');
+import AcsUserModel = require('../models/ACS/acsUserModel');
+import FileModel = require('../models/ACS/fileModel');
+import FolderModel = require('../models/ACS/folderModel');
 
-import TestConfig = require('../../test.config');
-import Util = require('../../util/util');
+import TestConfig = require('../test.config');
+import Util = require('../util/util');
 
 import AlfrescoApi = require('alfresco-js-api-node');
-import { UploadActions } from '../../actions/ACS/upload.actions';
+import { UploadActions } from '../actions/ACS/upload.actions';
 
 describe('Search component - Search Bar', () => {
