+++ e2e/search/search_multiselect.e2e.ts
@@ -15,22 +15,22 @@
  * limitations under the License.
  */
 
-import TestConfig = require('../../test.config');
+import TestConfig = require('../test.config');
 import AlfrescoApi = require('alfresco-js-api-node');
 
-import Util = require('../../util/util');
-import resources = require('../../util/resources');
-import CONSTANTS = require('../../util/constants');
-import { UploadActions } from '../../actions/ACS/upload.actions';
+import Util = require('../util/util');
+import resources = require('../util/resources');
+import CONSTANTS = require('../util/constants');
+import { UploadActions } from '../actions/ACS/upload.actions';
 import { browser } from 'protractor';
 
-import { LoginPage } from '../../pages/adf/loginPage';
-import { SearchDialog } from '../../pages/adf/dialog/searchDialog';
-import { SearchResultsPage } from '../../pages/adf/searchResultsPage';
-import { SearchFiltersPage } from '../../pages/adf/searchFiltersPage';
+import { LoginPage } from '../pages/adf/loginPage';
+import { SearchDialog } from '../pages/adf/dialog/searchDialog';
+import { SearchResultsPage } from '../pages/adf/searchResultsPage';
+import { SearchFiltersPage } from '../pages/adf/searchFiltersPage';
 
-import AcsUserModel = require('../../models/ACS/acsUserModel');
-import FileModel = require('../../models/ACS/fileModel');
+import AcsUserModel = require('../models/ACS/acsUserModel');
+import FileModel = require('../models/ACS/fileModel');
 
 describe('Search Component - Multi-Select Facet', () => {
     let loginPage = new LoginPage();
